# usm: gotls: performance improvement