--- a/pkg/network/go/bininspect/types.go
+++ b/pkg/network/go/bininspect/types.go
@@ -54,13 +54,12 @@ type elfMetadata struct {
 
 // Result is the result of the binary inspection process.
 type Result struct {
-	Arch                 GoArch
-	ABI                  GoABI
-	GoVersion            goversion.GoVersion
-	IncludesDebugSymbols bool
-	Functions            map[string]FunctionMetadata
-	StructOffsets        map[FieldIdentifier]uint64
-	GoroutineIDMetadata  GoroutineIDMetadata
+	Arch                GoArch
+	ABI                 GoABI
+	GoVersion           goversion.GoVersion
+	Functions           map[string]FunctionMetadata
+	StructOffsets       map[FieldIdentifier]uint64
+	GoroutineIDMetadata GoroutineIDMetadata
 }
 
 // GoArch only includes supported architectures,