<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Draft PRs should be prefixed with `[WIP]` in their title.

-->
### What does this PR do?

Major refactor in the goTLS inspection code.
The changes are divied into 2 changes:
1. Reimplementing GetAllSymbolsByName to be more effificent in memory and runtime. The new version is getting the list of wanted symbols to fetch, and returns the relevant symbols only.
2. Removing irrelevant call to HasDwarfInfo which allocated a lot of memory without any purpose

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

### Motivation
Solve memory and runtime overhead

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

Run 2 benchmarks on `main` and this branch.
The first benchmark tried to inpsect `docker` binary which has the relevant symbols, and trying to inpsect the test binary which has many symbols but none are tls.

on `main` branch 

```
BenchmarkGetSymbolsByNameDocker-16    	      10	 100817750 ns/op	132288267 B/op	  530289 allocs/op
BenchmarkGetSymbolsByNameTest-16     	      25	  43744169 ns/op	22586734 B/op	   63665 allocs/op

BenchmarkGetSymbolsByNameDocker-16    	      14	  83114233 ns/op	132287389 B/op	  530286 allocs/op
BenchmarkGetSymbolsByNameTest-16     	      28	  39570493 ns/op	22586641 B/op	   63664 allocs/op
```

On this branch

```
BenchmarkGetSymbolsByNameDocker-16    	     334	   3493458 ns/op	   93019 B/op	    1037 allocs/op
BenchmarkGetSymbolsByNameTest-16     	     475	   2537140 ns/op	   88113 B/op	     861 allocs/op

BenchmarkGetSymbolsByNameDocker-16    	     312	   3520427 ns/op	   93015 B/op	    1037 allocs/op
BenchmarkGetSymbolsByNameTest-16     	     470	   2588324 ns/op	   88120 B/op	     861 allocs/op
```
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.